# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for GuruEOB5 - school management app for teachers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type TeacherRole = typeof TeacherRole[keyof typeof TeacherRole];


export const TeacherRole = {
  admin: 'admin',
  guru: 'guru',
} as const;

export interface Teacher {
  id: string;
  username: string;
  name: string;
  role: TeacherRole;
  school: string;
  createdAt: string;
}

export interface RegisterInput {
  username: string;
  password: string;
  name: string;
  school: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UpdateTeacherInputRole = typeof UpdateTeacherInputRole[keyof typeof UpdateTeacherInputRole];


export const UpdateTeacherInputRole = {
  admin: 'admin',
  guru: 'guru',
} as const;

export interface UpdateTeacherInput {
  name?: string;
  role?: UpdateTeacherInputRole;
  school?: string;
}

export type StudentJenisKelamin = typeof StudentJenisKelamin[keyof typeof StudentJenisKelamin];


export const StudentJenisKelamin = {
  L: 'L',
  P: 'P',
} as const;

export interface Student {
  id: string;
  nisn: string;
  namaLengkap: string;
  kelas: string;
  jenisKelamin: StudentJenisKelamin;
  school: string;
  createdAt: string;
}

export type StudentInputJenisKelamin = typeof StudentInputJenisKelamin[keyof typeof StudentInputJenisKelamin];


export const StudentInputJenisKelamin = {
  L: 'L',
  P: 'P',
} as const;

export interface StudentInput {
  nisn?: string;
  namaLengkap: string;
  kelas: string;
  jenisKelamin: StudentInputJenisKelamin;
  school: string;
}

export interface Subject {
  id: string;
  name: string;
  teacherId: string;
  createdAt: string;
}

export interface SubjectInput {
  name: string;
  teacherId: string;
}

export interface AdminDocument {
  id: string;
  subjectId: string;
  name: string;
  description?: string;
  uploadedAt: string;
}

export interface AdminDocumentInput {
  subjectId: string;
  name: string;
  description?: string;
}

export interface JournalEntry {
  id: string;
  subjectId: string;
  teacherId: string;
  tanggal: string;
  kelas: string;
  materi: string;
  catatan?: string;
  createdAt: string;
}

export interface JournalEntryInput {
  subjectId: string;
  tanggal: string;
  kelas: string;
  materi: string;
  catatan?: string;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  hadir: 'hadir',
  izin: 'izin',
  sakit: 'sakit',
  alpa: 'alpa',
} as const;

export interface AttendanceRecord {
  id: string;
  studentId: string;
  subjectId: string;
  tanggal: string;
  status: AttendanceRecordStatus;
  createdAt: string;
}

export type AttendanceRecordInputStatus = typeof AttendanceRecordInputStatus[keyof typeof AttendanceRecordInputStatus];


export const AttendanceRecordInputStatus = {
  hadir: 'hadir',
  izin: 'izin',
  sakit: 'sakit',
  alpa: 'alpa',
} as const;

export interface AttendanceRecordInput {
  studentId: string;
  subjectId: string;
  tanggal: string;
  status: AttendanceRecordInputStatus;
}

export type GradeJenis = typeof GradeJenis[keyof typeof GradeJenis];


export const GradeJenis = {
  tugas: 'tugas',
  uts: 'uts',
  uas: 'uas',
} as const;

export interface Grade {
  id: string;
  studentId: string;
  subjectId: string;
  jenis: GradeJenis;
  nilai: number;
  tanggal: string;
  createdAt: string;
}

export type GradeInputJenis = typeof GradeInputJenis[keyof typeof GradeInputJenis];


export const GradeInputJenis = {
  tugas: 'tugas',
  uts: 'uts',
  uas: 'uas',
} as const;

export interface GradeInput {
  studentId: string;
  subjectId: string;
  jenis: GradeInputJenis;
  nilai: number;
  tanggal: string;
}

export type PointRecordJenis = typeof PointRecordJenis[keyof typeof PointRecordJenis];


export const PointRecordJenis = {
  positif: 'positif',
  negatif: 'negatif',
} as const;

export interface PointRecord {
  id: string;
  studentId: string;
  jenis: PointRecordJenis;
  poin: number;
  keterangan: string;
  tanggal: string;
  createdAt: string;
}

export type PointRecordInputJenis = typeof PointRecordInputJenis[keyof typeof PointRecordInputJenis];


export const PointRecordInputJenis = {
  positif: 'positif',
  negatif: 'negatif',
} as const;

export interface PointRecordInput {
  studentId: string;
  jenis: PointRecordInputJenis;
  poin: number;
  keterangan: string;
  tanggal: string;
}

export type DashboardSummaryProgresJurnalBulanIniItem = {
  minggu: string;
  jumlah: number;
};

export interface DashboardSummary {
  totalSiswa: number;
  totalGuru: number;
  totalDokumen: number;
  jurnalHariIniTerisi: boolean;
  progresJurnalBulanIni: DashboardSummaryProgresJurnalBulanIniItem[];
  kelengkapanAdministrasiPersen: number;
  schoolName: string;
  tahunAjaran: string;
  semester: string;
}

export type ListStudentsParams = {
search?: string;
};

export type ListDocumentsParams = {
subjectId?: string;
};

export type ListJournalEntriesParams = {
subjectId?: string;
};

export type ListAttendanceParams = {
subjectId?: string;
date?: string;
};

export type ListGradesParams = {
subjectId?: string;
studentId?: string;
};

export type ListPointsParams = {
studentId?: string;
};