# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for GuruEOB5 - school management app for teachers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new teacher account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "200":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List administrative documents
      parameters:
        - name: subjectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDocument"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create an administrative document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDocumentInput"
      responses:
        "200":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDocument"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List teaching journal entries
      parameters:
        - name: subjectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a teaching journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "200":
          description: Created journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: subjectId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendanceRecord
      tags: [attendance]
      summary: Create/record attendance for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRecordInput"
      responses:
        "200":
          description: Created attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: List grades
      parameters:
        - name: subjectId
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: Record a grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "200":
          description: Created grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /points:
    get:
      operationId: listPoints
      tags: [points]
      summary: List student points
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of point records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PointRecord"
    post:
      operationId: createPoint
      tags: [points]
      summary: Record a point entry for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointRecordInput"
      responses:
        "200":
          description: Created point record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointRecord"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    Teacher:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, guru]
        school:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, username, name, role, school, createdAt]
    RegisterInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        school:
          type: string
      required: [username, password, name, school]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
...
    Student:
      type: object
      properties:
        id:
          type: string
        nisn:
          type: string
        namaLengkap:
          type: string
        kelas:
          type: string
        jenisKelamin:
          type: string
          enum: [L, P]
        school:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, nisn, namaLengkap, kelas, jenisKelamin, school, createdAt]
    StudentInput:
      type: object
      properties:
        nisn:
          type: string
        namaLengkap:
          type: string
        kelas:
          type: string
        jenisKelamin:
          type: string
          enum: [L, P]
        school:
          type: string
      required: [namaLengkap, kelas, jenisKelamin, school]
...
    AdminDocument:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        name:
          type: string
        description:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required: [id, subjectId, name, uploadedAt]
    AdminDocumentInput:
      type: object
      properties:
        subjectId:
          type: string
        name:
          type: string
        description:
          type: string
      required: [subjectId, name]
    JournalEntry:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        teacherId:
          type: string
        tanggal:
          type: string
        kelas:
          type: string
        materi:
          type: string
        catatan:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, subjectId, teacherId, tanggal, kelas, materi, createdAt]
    JournalEntryInput:
      type: object
      properties:
        subjectId:
          type: string
        tanggal:
          type: string
        kelas:
          type: string
        materi:
          type: string
        catatan:
          type: string
      required: [subjectId, tanggal, kelas, materi]
    AttendanceRecord:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        subjectId:
          type: string
        tanggal:
          type: string
        status:
          type: string
          enum: [hadir, izin, sakit, alpa]
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, subjectId, tanggal, status, createdAt]
    AttendanceRecordInput:
      type: object
      properties:
        studentId:
          type: string
        subjectId:
          type: string
        tanggal:
          type: string
        status:
          type: string
          enum: [hadir, izin, sakit, alpa]
      required: [studentId, subjectId, tanggal, status]
    Grade:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        subjectId:
          type: string
        jenis:
          type: string
          enum: [tugas, uts, uas]
        nilai:
          type: number
        tanggal:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, subjectId, jenis, nilai, tanggal, createdAt]
    GradeInput:
      type: object
      properties:
        studentId:
          type: string
        subjectId:
          type: string
        jenis:
          type: string
          enum: [tugas, uts, uas]
        nilai:
          type: number
        tanggal:
          type: string
      required: [studentId, subjectId, jenis, nilai, tanggal]
    PointRecord:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        jenis:
          type: string
          enum: [positif, negatif]
        poin:
          type: number
        keterangan:
          type: string
        tanggal:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, jenis, poin, keterangan, tanggal, createdAt]
    PointRecordInput:
      type: object
      properties:
        studentId:
          type: string
        jenis:
          type: string
          enum: [positif, negatif]
        poin:
          type: number
        keterangan:
          type: string
        tanggal:
          type: string
      required: [studentId, jenis, poin, keterangan, tanggal]
...