# lib/api-zod/src/generated/types/adminDocument.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for GuruEOB5 - school management app for teachers
 * OpenAPI spec version: 0.1.0
 */

export interface AdminDocument {
  id: string;
  subjectId: string;
  name: string;
  description?: string;
  uploadedAt: Date;
}