# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for GuruEOB5 - school management app for teachers
 * OpenAPI spec version: 0.1.0
 */
import type { StudentJenisKelamin } from './studentJenisKelamin';

export interface Student {
  id: string;
  nisn: string;
  namaLengkap: string;
  kelas: string;
  jenisKelamin: StudentJenisKelamin;
  school: string;
  createdAt: Date;
}