# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for GuruEOB5 - school management app for teachers
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}